use mtree.get_proof and get rid of merkleroot's proof argument
raiden.channel.BalanceProof.get_known_unlocks should probably create a
mtree.Merkletree object and pass that to compute_proof_for_lock as it
creates multiple proofs with the same merkletree.

Some of the mtree tests could also be cleaned up. A bunch of them now
calls get_proof twice and compares the result. But it's too late for me
to fix that now.