Create release tooling #270
This allows us to create proof of concept releases in as few steps as possible:
On the master branch call

    ./prepare_poc_release.sh

Which will create a new PR. After the PR got merged, call:

    git checkout master
    git pull --rebase
    git tag <RELEASE_TAG>
    git push --tags

As a convention I propose that we release all PoC-versions in the `0.0.x` series,
where `x` equals the PoC number. So we should probably switch to `v0.0.x` tags
from the `PoC-x` tags.  Since pypi published versions should already follow this
standard, I believe this inconsistency with our prior PoC releases is okay.